highload_project/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
blinker==1.8.2
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
Django==5.1.3
django-allauth==65.2.0
django-celery-results==2.5.1
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-rest-framework==0.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
flower==2.0.1
gevent==24.10.3
geventhttpclient==2.3.1
greenlet==3.1.1
humanize==4.11.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
locust==2.32.1
MarkupSafe==3.0.2
msgpack==1.1.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
PyJWT==1.7.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
qrcode==7.4.2
requests==2.32.3
setuptools==75.3.0
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.2
zope.event==5.0
zope.interface==7.1.1
